# Don't store Stepping State or Breakpoint Rules in the context