--- a/src/MagneticReadHead.jl
+++ b/src/MagneticReadHead.jl
@@ -37,7 +37,7 @@ function iron_debug(debugbody, ctx)
         nothing
     finally
         # Disable any stepping left-over
-        ctx.metadata.stepping_mode =  StepContinue
+        GLOBAL_STEPPING_MODE[] =  StepContinue
     end
 end
 
@@ -47,7 +47,7 @@ Run until the_code until a breakpoint is hit.
 """
 macro run(body)
     quote
-        ctx = HandEvalCtx($(__module__), StepContinue)
+        ctx = HandEvalCtx($(__module__))
         iron_debug(ctx) do
             $(esc(body))
         end
@@ -65,7 +65,7 @@ macro enter(body)
         body = MacroTools.striplines(body)
         break_target = :(InteractiveUtils.which($(body.args[1]), Base.typesof($(body.args[2:end])...)))
         quote
-            ctx = HandEvalCtx($(__module__), StepContinue)
+            ctx = HandEvalCtx($(__module__))
             set_breakpoint!($(esc(break_target)))
             try
                 iron_debug(ctx) do