--- a/src/break_action.jl
+++ b/src/break_action.jl
@@ -54,14 +54,14 @@ end
 # this function exists only for mocking so we can test it.
 breakpoint_hit(meth, statement_ind, variables) = nothing
 
-function iron_repl(metadata::HandEvalMeta, meth, statement_ind, variables)
+function iron_repl(meth, statement_ind, variables)
     breadcrumbs(meth, statement_ind)
 
     printstyled("Vars: "; color=:light_yellow)
     println(join(keys(variables), ", "))
     print_commands()
 
-    run_repl(variables, metadata.eval_module)
+    run_repl(variables, Main)
 end
 
 """
@@ -79,13 +79,12 @@ end
 
 What to do when a breakpoint is hit
 """
-function break_action(ctx, meth, statement_ind, slotnames, slotvals)
-    metadata = ctx.metadata
+function break_action(meth, statement_ind, slotnames, slotvals)
 
     variables = LittleDict(slotnames, slotvals)
     pop!(variables, Symbol("#self#"))
     breakpoint_hit(meth, statement_ind, variables)
 
-    code_word = iron_repl(metadata, meth, statement_ind, variables)
+    code_word = iron_repl(meth, statement_ind, variables)
     actions[code_word].act()
 end